agents-cloud-run/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pyyaml==6.0.1
httpx==0.25.2
python-multipart==0.0.6
google-cloud-secret-manager==2.18.0
google-cloud-logging==3.8.0

